Properly handle temperature units on climate (#43)
ESPHome always uses celsius - which is fine if your system is in celsius.
But if you want fahrenheit, we need to make sure to tag the incoming units
as celsius so that openHAB can do the conversion.

On the command side, we need to do the conversion ourselves (we might
receive the command in either Celsius or Fahrenheit)

Signed-off-by: Cody Cutrer <cody@cutrer.us>